Experimental Ship Controls
Base off the Helicopter Script from the Unity Asset Store, I tried to
recreate a script that is similar for our Submarine Ship. The Script is
still in progress, as there are some bug and error in the script. An
Update of the file will be posted later. With that Noting on the
Controls script does not currently work.