Jose's assets were still being a butt
Jose's assets were causing problems due to Jose's branch being deleted
and a new branch being created. The problem was there was already a Jose
folder in the master and Jose made a new Jose folder in his new branch,
causing github to become confused about which Jose folder was the real
one.