Dossier Switching
Got to work on the Canvas System for the games Console Readings. Also
implemented a script that makes use of the 'P' button to switch between
"pages" of the dossier. Need to work on camera seeing, seeing about
making the shit with perspective a little less apparent and a little
more proper for visual representation.