BlurOxyDown Script Update
BlurOxyDown script was updated for the user of the script can change the
value of to their personal use. There are tooltip that explain each
usage. There are also header to separate the variables. The changes I
made as Follow: There is now a percentage slider to change how much
percentage of oxygen will start taking in affect. There is also a A
minimum Salutation value that the it will go down to instead of 0
(unless they want to you can). I stand by the the debugging variables
because they were my purpose.