Oxygen sound effect
I started to implement the sound effect with the oxygen going down. It
is a work in progress that works with Unity Audio Mixer. Right now you
can can control the volume with a slider, but not when it drops down to
a certain percentage. it works, but it is not subtracting correctly.